Fixed up long constants and shifts that needed to be cast to correct type.
Copied from Perforce
 Change: 178095
 ServerID: perforce.ravenbrook.com

+++ mps/code/table.c
@@ -157,7 +157,7 @@ extern Res TableCreate(Table *tableReturn, size_t length)
   if(table == NULL) goto failMallocTable;
   if (length < 2) length = 2;
   /* Table size is length rounded up to the next power of 2. */
-  table->length = 1 << (sizeFloorLog2(length-1) + 1);
+  table->length = (size_t)1 << (sizeFloorLog2(length-1) + 1);
   table->count = 0;
   table->limit = (size_t)(.5 * length);
   table->array = malloc(sizeof(TableEntryStruct) * length);
